% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{combine_data_efficient}
\alias{combine_data_efficient}
\title{Robust data combination with pre-allocation}
\usage{
combine_data_efficient(data_list, initial_size = 1000)
}
\arguments{
\item{data_list}{List of data vectors to combine}

\item{initial_size}{Initial pre-allocation size}
}
\value{
Combined vector
}
\description{
Robust data combination with pre-allocation
}
