% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_example_data}
\alias{create_example_data}
\title{Create example dataset for testing}
\usage{
create_example_data(n_datasets = 10, n_points = 1000)
}
\description{
Create example dataset for testing
}
