% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{print_analysis_summary}
\alias{print_analysis_summary}
\title{Print a nice summary of analysis results}
\usage{
print_analysis_summary(results)
}
\description{
Print a nice summary of analysis results
}
