% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_functions.R
\name{print_reliability_summary}
\alias{print_reliability_summary}
\title{Print a nice summary of reliability analysis results}
\usage{
print_reliability_summary(window_size, metrics)
}
\description{
Print a nice summary of reliability analysis results
}
