% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_functions.R
\name{save_reliability_results}
\alias{save_reliability_results}
\title{Save reliability analysis results}
\usage{
save_reliability_results(results, output_name = "reliability_analysis")
}
\description{
Save reliability analysis results
}
