% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{summarize_datasets_optimized}
\alias{summarize_datasets_optimized}
\title{Optimized summary statistics calculation}
\usage{
summarize_datasets_optimized(datasets)
}
\arguments{
\item{datasets}{List of datasets}
}
\value{
Summary statistics data frame
}
\description{
Optimized summary statistics calculation
}
