% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{validate_data_format}
\alias{validate_data_format}
\title{Check if data is in the correct format for analysis}
\usage{
validate_data_format(datasets)
}
\description{
Check if data is in the correct format for analysis
}
